fix: virtualize networks list for added performance
