+++ apps/extension/src/ui/apps/dashboard/routes/Networks/analytics.ts
@@ -1,0 +1,8 @@
+import { AnalyticsPage } from "@ui/api/analytics"
+
+export const ANALYTICS_PAGE: AnalyticsPage = {
+  container: "Fullscreen",
+  feature: "Settings",
+  featureVersion: 1,
+  page: "Settings - Networks",
+}
