+++ apps/extension/src/ui/domains/Settings/ManageNetworks/NetworkForm/SubNetworkForm.tsx
@@ -2,9 +2,9 @@ import { RequestUpsertCustomChain } from "@core/domains/chains/types"
 import { CustomNativeToken } from "@core/domains/tokens/types"
 import { yupResolver } from "@hookform/resolvers/yup"
 import { HeaderBlock } from "@talisman/components/HeaderBlock"
-import { ArrowRightIcon } from "@talisman/theme/icons"
 import { shortenAddress } from "@talisman/util/shortenAddress"
 import { Chain, ChainId, CustomChain } from "@talismn/chaindata-provider"
+import { ArrowRightIcon } from "@talismn/icons"
 import { classNames } from "@talismn/util"
 import { useQuery } from "@tanstack/react-query"
 import { api } from "@ui/api"
